ref: callable based permission handling now replaces the current permission
